requirements.txt
fastapi
uvicorn
gunicorn
httpx
pydantic
pydantic-settings
python-multipart
openai
anthropic
numpy
python-dotenv
sentence-transformers>=2.2.2
torch
scikit-learn>=1.3.0
requests
beautifulsoup4
psycopg2-binary
google-api-python-client
google-auth
typing-extensions>=4.0.0
